requirements.txt
binance-connector==3.8.1
binance-futures-connector==4.0.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
humanize==4.10.0
idna==3.7
numpy==2.0.1
pillow==10.4.0
pyaes==1.6.1
pyasn1==0.6.0
pycryptodome==3.20.0
pyodbc==5.1.0
python-dotenv==1.0.1
reportlab==4.2.2
requests==2.32.3
rsa==4.9
Telethon==1.36.0
urllib3==2.2.2
websocket-client==1.8.0
